PR comment fixes 1
