+++ src/dotnet-scaffolding/Microsoft.DotNet.Scaffolding.CodeModification/Helpers/ProjectModifierHelper.cs
@@ -461,7 +461,7 @@ internal static class ProjectModifierHelper
     /// <param name="fileDoc"></param>
     /// <param name="codeChanges"></param>
     /// <returns>updated document, or null if no changes made</returns>
-    internal static async Task<TextDocument?> ModifyDocumentTextAsync(TextDocument? fileDoc, IEnumerable<CodeSnippet>? codeChanges)
+    internal static async Task<T?> ModifyDocumentTextAsync<T>(T? fileDoc, IEnumerable<CodeSnippet>? codeChanges) where T : TextDocument
     {
         if (fileDoc is null || codeChanges is null || !codeChanges.Any())
         {
@@ -505,7 +505,7 @@ internal static class ProjectModifierHelper
         var solution = fileDoc.Project.Solution;
         var updatedSolution = solution.WithAdditionalDocumentText(fileDoc.Id, updatedSourceText);
         var updatedTextDocument = updatedSolution.GetAdditionalDocument(fileDoc.Id);
-        return updatedTextDocument;
+        return updatedTextDocument as T;
     }
 
     internal static async Task UpdateDocument(Document document)
